+++ changelog.md
@@ -1,4 +1,140 @@
 # Changelog
+
+## [1.1.18] - 2025-11-24
+### Ajouté
+- Gestion de la mise en sitemap des pages de Glossaire et de Metadatas
+- Gestion du cache et de l'exclusion dans WP-ROCKET
+- Gestion des rewrites proprement et suppression des soucis de Permalinks
+- Gestion des insertions de Elaia dans les sites
+- Ajout de sécurité
+
+### Modifié
+- Restructuration complète du code source
+
+### Supprimé
+- N/A
+
+
+## [1.1.17] - 2025-11-06
+### Ajouté
+- Rien
+
+### Modifié
+- Correction height 
+
+### Supprimé
+- Rien pour cette version
+
+
+## [1.1.16] - 2025-11-06
+### Ajouté
+- Rien
+
+### Modifié
+- Correction z-index
+
+### Supprimé
+- Rien pour cette version
+
+
+## [1.1.15] - 2025-10-09
+### Ajouté
+- Cache le chatbot si form ouvert
+
+### Modifié
+- Bug corrigé sur placeholder
+
+### Supprimé
+- Rien pour cette version
+
+
+## [1.1.14] - 2025-10-09
+### Ajouté
+- Cache le chatbot si form ouvert
+
+### Modifié
+- Bug corrigé sur placeholder
+
+### Supprimé
+- Rien pour cette version
+
+## [1.1.13] - 2025-10-08
+### Ajouté
+- Ajout du geo pour metadatas
+
+### Modifié
+- !important sur le design faq
+
+### Supprimé
+- Rien pour cette version
+
+## [1.1.12] - 2025-10-06
+### Ajouté
+- Ajout page elaia-metadata
+
+### Modifié
+- Rien pour cette version
+
+### Supprimé
+- Rien pour cette version
+
+## [1.1.11] - 2025-09-30
+### Ajouté
+- Ajout page elaia-faq
+
+### Modifié
+- Rien pour cette version
+
+### Supprimé
+- Rien pour cette version
+
+
+## [1.1.8] - 2025-09-22
+### Ajouté
+- Balise script pour encart overview
+
+### Modifié
+- Rien pour cette version
+
+### Supprimé
+- Rien pour cette version
+
+## [1.1.7] - 2025-07-28
+### Ajouté
+- Rien pour cette version
+
+### Modifié
+- L'ajout de la balise script
+
+### Supprimé
+- Rien pour cette version
+
+
+
+## [1.1.6] - 2025-07-18
+### Ajouté
+- Rien pour cette version
+
+### Modifié
+- Rien pour cette version
+
+### Supprimé
+- Checker et mettre à jours les plugins
+- Les logs 
+
+
+## [1.1.5] - 2025-07-11
+### Ajouté
+- Ajout de la langue du site dans l'url pour appeler le chatbot
+
+### Modifié
+- L'url pour appeler le chatbot
+
+### Supprimé
+- Rien pour cette version
+
+
+# Changelog
 ## [1.1.4] - 2025-07-01
 ### Ajouté
 - Ajout d’un système de vérification des mises à jour pour permettre la mise à jour automatique du plugin